+++ .github/workflows/windows.yml
@@ -30,22 +30,4 @@ jobs:
           python -m pip install -e .
       - name: Test with pytest
         run: |
-          python -m pytest -vs src/exoplanet
-
-#       - name: Set up Python 3.8
-#         uses: actions/setup-python@v1
-#         with:
-#           python-version: 3.8
-#       - name: Install dependencies
-#         run: |
-#           python -m pip install --upgrade pip
-#           python -m pip install .
-#       - name: Lint with black and mypy
-#         run: |
-#           python -m pip install black mypy
-#           black --check src
-#           mypy --ignore-missing-imports src
-#       - name: Test with pytest
-#         run: |
-#           python -m pip install pytest pytest-aiohttp
-#           python -m pytest -vs tests
+          python -m pytest -vs tests
